Only install /etc/init.d/rc.d/* if systemd-sysv-install is not present

+++ rpms/SPECS/wazuh-manager.spec
@@ -74,14 +74,17 @@ echo 'USER_CREATE_SSL_CERT="n"' >> ./etc/preloaded-vars.conf
 ./install.sh
 
 # Create directories
-mkdir -p ${RPM_BUILD_ROOT}%{_initrddir}
+
+if [ ! -f /usr/lib/systemd/systemd-sysv-install ]; then mkdir -p ${RPM_BUILD_ROOT}%{_initrddir}; fi
 mkdir -p ${RPM_BUILD_ROOT}%{_localstatedir}/.ssh
 
 # Copy the installed files into RPM_BUILD_ROOT directory
 cp -pr %{_localstatedir}/* ${RPM_BUILD_ROOT}%{_localstatedir}/
 mkdir -p ${RPM_BUILD_ROOT}/usr/lib/systemd/system/
-sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/ossec-hids-rh.init
-install -m 0755 src/init/templates/ossec-hids-rh.init ${RPM_BUILD_ROOT}%{_initrddir}/wazuh-manager
+if [ ! -f /usr/lib/systemd/systemd-sysv-install ]; then
+  sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/ossec-hids-rh.init
+  install -m 0755 src/init/templates/ossec-hids-rh.init ${RPM_BUILD_ROOT}%{_initrddir}/wazuh-manager
+fi
 sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/wazuh-manager.service
 install -m 0644 src/init/templates/wazuh-manager.service ${RPM_BUILD_ROOT}/usr/lib/systemd/system/
 
@@ -326,16 +329,9 @@ fi
 
   if [ -n "$sles" ] && [ $(ps --no-headers -o comm 1) == "systemd" ]; then
     if [ -f /etc/init.d/wazuh-manager ]; then
-      rm -f /etc/init.d/wazuh-manager
+      rm -f %{_initrddir}//wazuh-manager
     fi
   fi
-
-if [ -f /etc/os-release ]; then
-  source /etc/os-release
-  if (( [ "${NAME}" = "Red Hat Enterprise Linux" ] || [ "${NAME}" = "Rocky Linux" ] || [ "${NAME}" = "AlmaLinux" ] || [ "CentOS Stream" ]) && [ "$((${VERSION_ID:0:1}))" -ge 9 ]) || ( [ "${NAME}" = "Fedora" ] && [ "$((${VERSION_ID}))" -ge 34 ] ); then
-    rm -f %{_initrddir}/wazuh-manager
-  fi
-fi
 
 # Generation auto-signed certificate if not exists
 if [ ! -f "%{_localstatedir}/etc/sslmanager.key" ] && [ ! -f "%{_localstatedir}/etc/sslmanager.cert" ]; then
